refactor ErrorMessage export
