+++ src/frontend/components/shared/index.js
@@ -9,3 +9,4 @@ export * from './Scrollbar'
 export * from './Circle'
 export * from './LaundryOverview'
 export * from './NoData'
+export * from './ErrorMessage'
